Fixes to compile on Linux w/ gcc and silence warnings (#12)

+++ src/tests/test.c
@@ -137,7 +137,7 @@ test_rand64()
     uint64_t random = 0;
 
     random = con4m_rand64();
-    printf("Random value: %16llx\n", random);
+    printf("Random value: %16llx\n", (unsigned long long)random);
     assert(random != 0);
 }
 
@@ -505,7 +505,7 @@ main(int argc, char **argv, char **envp)
         test4();
         table_test();
 
-        printf("Sample style: %.16llx\n", style1);
+        printf("Sample style: %.16llx\n", (unsigned long long)style1);
         sha_test();
 
         type_tests();
@@ -550,6 +550,6 @@ main(int argc, char **argv, char **envp)
         stream_putc(sout, '\n');
 
         bottom = top + q;
-        printf("(start) = %p; (end) = %p (%llu bytes)\n", (void *)top, (void *)bottom, q);
+        printf("(start) = %p; (end) = %p (%llu bytes)\n", (void *)top, (void *)bottom, (unsigned long long)q);
     }
 }
